refactor: decompose MaintenanceClient into custom hooks (#74)
Extract 4 stateful logic clusters from the 723-line god component:

- useMaintenanceViewState: view mode, search, sort, filtered/sorted repos
- useRestoreDialog: restore dialog state, board loading
- useMaintenanceComments: comment CRUD with optimistic updates
- useMaintenanceNoteModal: note modal with lazy project info loading

MaintenanceClient.tsx reduced from 723 to 508 lines. Each hook is
independently testable and follows single-responsibility principle.
All 21 existing MaintenanceClient tests pass unchanged.
